search methods all consistent now
